Implement lexer
